enable reloading of saved parameter sets
If paramsets.Rdata file exists, load it in preference to the
defparamsets.Rdata file.  paramsets.Rdata is only present if the user
has first saved a parameter set by labelleing a parameter set 'saveAs'
field with a character string.  Enables more efficient setup of
processing when the user will be processing several datasets with
similar parameters.